requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pdfplumber
sentence-transformers
chromadb
requests